file

+++ src/main/java/com/example/spellingcheck/service/implement/FileService.java
@@ -20,7 +20,7 @@ public class FileService<T> implements IFileService {
     @Override
     public ResponseEntity<Resource> getFile(String filePath) {
         try {
-            Path path = Paths.get(Constants.ZEEK_CONFIG_PATH + filePath);
+            Path path = Paths.get(filePath);
             Resource resource = new UrlResource(path.toUri());
             File directory = resource.getFile();
             if (directory.isFile())
@@ -53,19 +53,24 @@ public class FileService<T> implements IFileService {
                             zeekConfigPath = Constants.ZEEK_CONFIG_PATH.replace('/', '\\');
                         String filePath = file.getPath();
                         String newPath = filePath.substring(filePath.indexOf(zeekConfigPath) + zeekConfigPath.length());
+                        newPath = newPath.replace('\\', '/');
 
                         if (file.isFile()) {
-                            System.out.println("File: " + file.getName());
                             json.append("{\"type\": \"file\", \"path\": \"").append(newPath).append("\", \"name\": \"").append(file.getName()).append("\"},");
                         } else if (file.isDirectory()) {
-                            System.out.println("Directory: " + file.getName());
                             json.append("{\"type\": \"directory\", \"path\": \"").append(newPath).append("\", \"name\": \"").append(file.getName()).append("\"},");
                         }
                     }
                 } else {
                     System.out.println("The directory is empty or an error occurred.");
                 }
-                json.append("]");
+                if (json.length() != 1) {
+                    json.deleteCharAt(json.length() - 1);
+                    json.append("]");
+                } else {
+                    json.append("]");
+                }
+
                 return ResponseEntity.ok(json.toString());
             } else {
                 System.out.println("The path is not a directory.");
@@ -77,7 +82,10 @@ public class FileService<T> implements IFileService {
     }
 
     @Override
-    public ResponseEntity<String> addFile(String path, String content) {
+    public ResponseEntity<String> addFile(String path, String content, boolean create) {
+        File file = new File(Constants.ZEEK_CONFIG_PATH + path);
+        if (file.exists() && !create)
+            return ResponseEntity.badRequest().body("file existed");
         try {
             // Tạo một đối tượng FileWriter để ghi nội dung vào tệp mới
             FileWriter fileWriter = new FileWriter(Constants.ZEEK_CONFIG_PATH + path);
